Add files via upload
Dashboard Type: Data Analysis and Visualization
• Tool Used: Tableau
• Purpose: Analyzing and presenting insights on the COVID-19 situation in India
• Project Overview: The COVID-19 Dashboard project aims to provide a comprehensive and interactive platform for analyzing and visualizing the COVID-19 
situation in India. This project leverages Tableau to create a user-friendly dashboard that offers real-time insights, trends, and key metrics related to the pandemic. 
The dashboard serves as a valuable resource for healthcare professionals, policymakers, researchers, and the general public to understand the evolving dynamics 
of COVID-19 in India.
• Features Used: 
• Live and Extract Data Connections: in tableau using live data connections of tableau.
• Combine data from various sources: such as databases, spreadsheets, and online sources to create a unified dataset (Kaggle).
• Interactive Visualizations: Charts and Graphs, Geographical Maps, Dashboards and Stories.